Refactor node classes into core.node package
This may allow for better "protected" vs "public" visibility.
