+++ src/main/java/ExpressionTree.java
@@ -1,5 +1,6 @@
 
 import core.*;
+import core.node.*;
 
 import java.math.BigDecimal;
 import java.util.Iterator;
