+++ src/main/java/core/ExpressionParser.java
@@ -4,6 +4,7 @@ import java.math.BigDecimal;
 import java.util.*;
 import java.util.regex.Pattern;
 
+import core.node.*;
 import exceptions.MissingInputException;
 
 /**
