+++ src/main/java/core/node/CalculusNode.java
@@ -1,4 +1,4 @@
-package core;
+package core.node;
 
 import exceptions.InvalidInputException;
