+++ src/main/java/core/node/NumNode.java
@@ -1,4 +1,4 @@
-package core;
+package core.node;
 
 import java.math.BigDecimal;
 import java.util.Map;
