+++ src/main/java/core/node/OperandNode.java
@@ -1,4 +1,4 @@
-package core;
+package core.node;
 
 /**
  * Represents the operands of the expression tree
