+++ src/main/java/core/node/OperationType.java
@@ -1,4 +1,4 @@
-package core;
+package core.node;
 
 public enum OperationType
 {
