+++ src/main/java/core/node/UnaryNode.java
@@ -1,4 +1,4 @@
-package core;
+package core.node;
 
 import util.tareknaj.BigFunctions;
