+++ src/main/java/core/node/VarNode.java
@@ -1,4 +1,4 @@
-package core;
+package core.node;
 
 import exceptions.UndefinedVariableException;
