+++ src/main/java/core/node/WallNode.java
@@ -1,4 +1,4 @@
-package core;
+package core.node;
 
 import java.math.BigDecimal;
 import java.util.Map;
